Transcript again
Before you approve this, please look at the page itself. [5:22] is not displayed on the page and is not in the page source. I tried to find a problem but all I found was an extra space after the colon, but there were other instances of that, so I don't know!